+++ src/types/assignActivityPage.ts
@@ -1,4 +1,4 @@
-interface ReservationTime {
+export interface ReservationTime {
   reservationDate: string; // 예약 날짜
   startTime: string; // 시작 시간
   endTime: string; // 종료 시간
